some modification in payment integration
