+++ backend/routes/payment.js
@@ -1,34 +1,11 @@
-
-const instance=require("../app.js");
+var express = require('express');
+var router = express.Router();
 const crypto =require("crypto")
 require("dotenv").config()
 
-
-exports.checkout=async (req,res)=>{
-  try{
-  const options={
-    amount:Number(req.body.price*100),
-    currency:"INR",
-  }
-  const order=await instance.orders.create(options);
-  console.log(order);
-  res.status(200).json({
-    success:true,
-    message:"order created successfully",
-    order,
-  })
-}
-catch(error){
-    console.log(error);
-    res.status(500).json({
-        success:false,
-        message:"failed in creating order"
-    })
- }
-};
-
-exports.paymentVerification=async (req,res)=>{
-   const {razorpay_order_id,
+router.get( "/",async (req,res)=>{
+   try {
+    const {razorpay_order_id,
           razorpay_payment_id,
           razorpay_signature
    }=req.body;
@@ -43,14 +20,23 @@ exports.paymentVerification=async (req,res)=>{
    
     const isAuthentic = expectedSignature === razorpay_signature;
     if(isAuthentic){
-       res.status(200).json({
+       return res.status(200).json({
         success:true,
         message:"payment is successfull"
        })
     }else{
-        res.status(500).json({
+       return res.status(400).json({
             success:false,
             message:"payment verification failed"
         });
     }
-};
+}catch(error){
+    console.log(error)
+    return res.status(500).json({
+        success:false,
+        message:"internal server error while payment verification"
+    })
+}
+});
+
+module.exports = router;
